backend/requirements.txt
accelerate==0.20.3
aiohttp==3.8.4
aiohttp-cors==0.7.0
aiorwlock==1.3.0
aiosignal==1.3.1
alembic==1.11.1
ansicon==1.89.0
anyio==3.7.1
appdirs==1.4.4
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
Authlib==1.2.1
azure-ai-formrecognizer==3.3.0b1
azure-common==1.1.28
azure-core==1.27.1
backoff==2.2.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
blessed==1.20.0
boilerpy3==1.0.6
boto3==1.28.1
botocore==1.31.1
cachetools==5.3.1
canals==0.2.2
cattrs==23.1.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.0.4
cloudpickle==2.2.1
colorama==0.4.6
coloredlogs==15.0.1
colorful==0.5.5
contourpy==1.1.0
cryptography==41.0.1
cycler==0.11.0
databricks-cli==0.17.7
datasets==2.13.1
decorator==5.1.1
dill==0.3.6
distlib==0.3.6
Django==4.2.3
django-cors-headers==4.1.0
dnspython==2.3.0
docker==6.1.3
docopt==0.6.2
elasticsearch==7.17.9
entrypoints==0.4
Events==0.4
exceptiongroup==1.1.2
faiss-cpu==1.7.2
farm-haystack==1.18.1
fastapi==0.100.0
ffmpeg-python==0.2.0
filelock==3.12.2
Flask==2.2.5
flatbuffers==23.5.26
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.1
google-auth==2.21.0
googleapis-common-protos==1.59.1
gpustat==1.1
greenlet==2.0.2
grpcio==1.43.0
h11==0.14.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
inflect==7.0.0
isodate==0.6.1
itsdangerous==2.1.2
jarowinkler==1.2.3
Jinja2==3.1.2
jinxed==1.2.0
jmespath==1.0.1
joblib==1.3.1
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
kiwisolver==1.4.4
langdetect==1.0.9
lazy-imports==0.3.1
llvmlite==0.40.1
loguru==0.7.0
lxml==4.9.3
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.2
mlflow==2.3.2
monotonic==1.6
more-itertools==9.1.0
mpmath==1.3.0
msgpack==1.0.5
msrest==0.7.1
multidict==6.0.4
multiprocess==0.70.14
networkx==3.1
nltk==3.8.1
num2words==0.5.12
numba==0.57.1
numpy==1.24.4
nvidia-ml-py==12.535.77
oauthlib==3.2.2
onnx==1.14.0
onnxruntime==1.15.1
onnxruntime-tools==1.7.0
openai-whisper==20230308
opencensus==0.11.2
opencensus-context==0.1.3
opensearch-py==2.2.0
outcome==1.2.0
packaging==23.1
pandas==2.0.3
pdf2image==1.16.3
Pillow==10.0.0
pinecone-client==2.2.2
pkgutil_resolve_name==1.3.10
platformdirs==3.8.1
posthog==3.0.1
prometheus-client==0.13.1
prompthub-py==4.0.0
protobuf==3.20.3
psutil==5.9.5
py-cpuinfo==9.0.0
py-spy==0.3.14
py3nvml==0.2.7
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.11
PyJWT==2.7.0
PyMuPDF==1.22.5
pyparsing==3.0.9
pyreadline3==3.4.1
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-frontmatter==1.0.0
python-magic-bin==0.4.14
pytz==2023.3
pywin32==306
PyYAML==6.0
quantulum3==0.9.0
querystring-parser==1.2.4
rank-bm25==0.2.2
rapidfuzz==2.7.0
ray==1.13.0
referencing==0.29.1
regex==2023.6.3
requests==2.28.2
requests-cache==0.9.8
requests-oauthlib==1.3.1
rpds-py==0.8.10
rsa==4.9
s3transfer==0.6.1
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.10.1
selenium==4.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
seqeval==1.2.2
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==1.4.49
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
sseclient-py==1.7.2
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
threadpoolctl==3.1.0
tika==2.6.0
tiktoken==0.4.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.30.1
trio==0.22.1
trio-websocket==0.10.3
typing_extensions==4.7.1
tzdata==2023.3
url-normalize==1.4.3
urllib3==1.26.16
uvicorn==0.16.0
validators==0.20.0
virtualenv==20.23.1
waitress==2.1.2
wcwidth==0.2.6
weaviate-client==3.18.0
webdriver-manager==3.8.6
websocket-client==1.6.1
Werkzeug==2.3.6
whitenoise==6.5.0
win32-setctime==1.1.0
wsproto==1.2.0
xmltodict==0.13.0
xxhash==3.2.0
yarl==1.9.2
zipp==3.16.0
